+++ CHANGE_LOG.md
@@ -2,6 +2,7 @@
 
 * Improved camera follow
 * Decreased play field from 150 to 75
+* Fixed world border distance check
 
 ## Version 0.2.4